+++ HISTORY.md
@@ -2,9 +2,11 @@
 ## Unreleased
 ### Public API Change
 * Iterator::SeekForPrev is now a pure virtual method. This is to prevent user who implement the Iterator interface fail to implement SeekForPrev by mistake.
+- Add `include_end` option to make the range end exclusive when `include_end == false` in `DeleteFilesInRange()`.
 
 ### New Features
 * Improve the performance of iterators doing long range scans by using readahead.
+- Add new function `DeleteFilesInRanges()` to delete files in multiple ranges at once for better performance.
 
 ### Bug Fixes
 * Fix `DisableFileDeletions()` followed by `GetSortedWalFiles()` to not return obsolete WAL files that `PurgeObsoleteFiles()` is going to delete.